requirements.txt
annotated-types==0.7.0
anyio==4.4.0
arrow==1.3.0
asgiref==3.8.1
binaryornot==0.4.4
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cookiecutter==2.6.0
distro==1.9.0
Django==5.0.7
django-browser-reload==1.13.0
django-tailwind==3.8.0
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
openai==1.37.1
packaging==24.1
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
text-unidecode==1.3
tqdm==4.66.4
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
